requirements.txt
# Core Dependencies
python-dotenv>=0.20.0
requests>=2.28.0
pandas>=2.0.0
numpy>=1.21.0

# Financial Data APIs
yfinance>=0.2.0
finnhub-python>=1.0.0
alpha-vantage>=2.3.0

# Machine Learning
scikit-learn>=1.0.0
xgboost>=1.7.0

# Sentiment Analysis
textblob>=0.17.0
nltk>=3.7
vader-sentiment>=3.3.2
praw>=7.6.0

# Broker Integration
alpaca-py>=0.15.0

# Visualization
matplotlib>=3.5.0
plotly>=5.0.0
streamlit>=1.20.0

# Utilities
tqdm>=4.60.0
pytest>=6.2.0

# Optional for advanced features
# lightgbm>=3.3.0
# tensorflow>=2.8.0
# torch>=1.10.0

